refactor(report-convert): move conversion functions into lib
