+++ bd-report-convert/src/main.rs
@@ -5,23 +5,11 @@
 // LICENSE file or at:
 // https://polyformproject.org/wp-content/uploads/2020/06/PolyForm-Shield-1.0.0.txt
 
-use std::ffi::{CStr, CString, c_char};
+use bd_report_convert::{bin_to_json, json_to_bin};
 use std::io::Write;
 use std::path::Path;
 use std::process::ExitCode;
-use std::{env, fs, slice};
-
-
-unsafe extern "C" {
-  fn bdrc_make_bin_from_json(
-    schema_data: *const u8,
-    json_data_path: *const c_char,
-    length_or_err: *mut i32,
-  ) -> *const u8;
-
-  fn bdrc_alloc_json(bin_data_path: *const c_char) -> *const c_char;
-  fn bdrc_json_free(json: *const c_char);
-}
+use std::{env, fs};
 
 fn print_usage(program: &str) -> ExitCode {
   eprintln!("converter for the bd issue report format\n");
@@ -45,57 +33,39 @@ pub fn main() -> Result<ExitCode, std::io::Error> {
     return Ok(ExitCode::from(74));
   }
   match (args.len(), action.as_str()) {
-    (3, "to-json") => print_json(input_path.as_str()),
+    (3, "to-json") => Ok(print_json(input_path.as_str())),
     (4, "to-bin") => write_bin(input_path.as_str(), &args[3]),
     _ => Ok(print_usage(&args[0])),
   }
 }
 
-fn print_json(input_path: &str) -> Result<ExitCode, std::io::Error> {
-  let data_path = CString::new(input_path)?;
-  let text = unsafe { bdrc_alloc_json(data_path.as_ptr()) };
-  let text_ptr = unsafe { CStr::from_ptr(text) };
-
-  match text_ptr.to_str() {
+fn print_json(input_path: &str) -> ExitCode {
+  match bin_to_json(input_path) {
     Ok(json) => {
       println!("{json}");
-      unsafe {
-        bdrc_json_free(text);
-      }
-      Ok(ExitCode::SUCCESS)
+      ExitCode::SUCCESS
     },
     Err(result) => {
       eprintln!("failed to convert file: {result}");
-      Ok(ExitCode::FAILURE)
+      ExitCode::FAILURE
     },
   }
 }
 
 fn write_bin(input_path: &str, output_path: &str) -> Result<ExitCode, std::io::Error> {
-  let schema_data = concat!(
-    include_str!("../../api/src/bitdrift_public/fbs/issue-reporting/v1/report.fbs"),
-    "\0"
-  );
-  let data_path = CString::new(input_path)?;
-  let mut length_or_err: i32 = 0;
-  let buf = unsafe {
-    bdrc_make_bin_from_json(
-      schema_data.as_bytes().as_ptr(),
-      data_path.as_ptr(),
-      &mut length_or_err,
-    )
-  };
-  if length_or_err > 0 {
-    let mut file = fs::OpenOptions::new()
-      .create(true)
-      .truncate(true)
-      .write(true)
-      .open(output_path)?;
-    let data = unsafe { slice::from_raw_parts(buf, length_or_err.unsigned_abs() as usize) };
-    file.write_all(data)?;
-    Ok(ExitCode::SUCCESS)
-  } else {
-    eprintln!("failed to create file: {length_or_err}");
-    Ok(ExitCode::from(74))
+  match json_to_bin(input_path) {
+    Ok(data) => {
+      let mut file = fs::OpenOptions::new()
+        .create(true)
+        .truncate(true)
+        .write(true)
+        .open(output_path)?;
+      file.write_all(data)?;
+      Ok(ExitCode::SUCCESS)
+    },
+    Err(code) => {
+      eprintln!("failed to create file: {code}");
+      Ok(ExitCode::from(74))
+    },
   }
 }
